Add files via upload
The project involves the real estate prices in the city of Ames located in Iowa State. Aim of this project is to create a prediction model for the real estate prices in coming months/years. It will also try to predict what price the residential house owners can sell their prices in current market scenario. This will help the people to exit the market in right time in such an inflated market and make correct decisions onto when they can sell their properties. It will also help new investors to make decision in order to what time will be correct for them to invest in the residential properties in Ames. I will be using Regression in order to create the prediction model for this given data set.